Move modules from `app` to `cardano-node`
