+++ nix/.stack.nix/cardano-node.nix
@@ -17,17 +17,37 @@
     components = {
       "library" = {
         depends = [
+          (hsPkgs.aeson)
+          (hsPkgs.async)
           (hsPkgs.base)
+          (hsPkgs.brick)
           (hsPkgs.bytestring)
           (hsPkgs.canonical-json)
           (hsPkgs.cborg)
+          (hsPkgs.containers)
+          (hsPkgs.contra-tracer)
+          (hsPkgs.cryptonite)
+          (hsPkgs.directory)
+          (hsPkgs.file-embed)
           (hsPkgs.formatting)
+          (hsPkgs.iproute)
+          (hsPkgs.iohk-monitoring)
+          (hsPkgs.io-sim-classes)
           (hsPkgs.mtl)
+          (hsPkgs.network)
           (hsPkgs.optparse-applicative)
+          (hsPkgs.process)
+          (hsPkgs.safe-exceptions)
+          (hsPkgs.serialise)
+          (hsPkgs.template-haskell)
           (hsPkgs.text)
           (hsPkgs.time)
+          (hsPkgs.typed-protocols)
+          (hsPkgs.typed-protocols-cbor)
+          (hsPkgs.vty)
           (hsPkgs.generic-monoid)
           (hsPkgs.cardano-binary)
+          (hsPkgs.cardano-crypto-class)
           (hsPkgs.cardano-crypto-wrapper)
           (hsPkgs.cardano-ledger)
           (hsPkgs.cardano-ledger-test)
@@ -43,42 +63,10 @@
         "cardano-node" = {
           depends = [
             (hsPkgs.base)
-            (hsPkgs.cardano-crypto-class)
-            (hsPkgs.cardano-crypto-wrapper)
-            (hsPkgs.cardano-ledger)
-            (hsPkgs.cardano-ledger-test)
             (hsPkgs.cardano-node)
             (hsPkgs.cardano-prelude)
             (hsPkgs.cardano-shell)
-            (hsPkgs.contra-tracer)
-            (hsPkgs.io-sim-classes)
-            (hsPkgs.iohk-monitoring)
-            (hsPkgs.ouroboros-network)
-            (hsPkgs.ouroboros-consensus)
-            (hsPkgs.typed-protocols-cbor)
-            (hsPkgs.aeson)
-            (hsPkgs.async)
-            (hsPkgs.bytestring)
-            (hsPkgs.cborg)
-            (hsPkgs.containers)
-            (hsPkgs.cryptonite)
-            (hsPkgs.directory)
-            (hsPkgs.file-embed)
-            (hsPkgs.formatting)
-            (hsPkgs.iproute)
-            (hsPkgs.lens)
-            (hsPkgs.mtl)
-            (hsPkgs.network)
             (hsPkgs.optparse-applicative)
-            (hsPkgs.process)
-            (hsPkgs.serialise)
-            (hsPkgs.safe-exceptions)
-            (hsPkgs.stm)
-            (hsPkgs.string-conv)
-            (hsPkgs.template-haskell)
-            (hsPkgs.text)
-            (hsPkgs.time)
-            (hsPkgs.typed-protocols)
             ] ++ (if system.isWindows
             then [ (hsPkgs.Win32) ]
             else [ (hsPkgs.unix) (hsPkgs.brick) (hsPkgs.vty) ]);
@@ -110,23 +98,17 @@
         "genesis-tool" = {
           depends = [
             (hsPkgs.base)
-            (hsPkgs.cardano-binary)
             (hsPkgs.cardano-crypto)
             (hsPkgs.cardano-crypto-wrapper)
             (hsPkgs.cardano-ledger)
             (hsPkgs.cardano-ledger-test)
             (hsPkgs.cardano-node)
             (hsPkgs.cardano-prelude)
-            (hsPkgs.contra-tracer)
             (hsPkgs.filepath)
             (hsPkgs.iohk-monitoring)
             (hsPkgs.pvss)
-            (hsPkgs.aeson)
-            (hsPkgs.aeson-pretty)
-            (hsPkgs.async)
             (hsPkgs.binary)
             (hsPkgs.bytestring)
-            (hsPkgs.canonical-json)
             (hsPkgs.cborg)
             (hsPkgs.containers)
             (hsPkgs.cryptonite)
